contrib/docker/: Add scripts for building a Docker image.
The previous Dockerfile used the "gcc" base image, resulting in a ~1.4
GByte Docker image. These script are building a static binary outside of
Docker and package that, resulting in a ~35 kByte image. It is using the
"Musl" libc because glibc's getaddrinfo(3) expects to be able to
load providers at runtime based on configuration in /etc/nsswitch.conf.